requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
transformers==4.36.2
torch==2.2.2
sentence-transformers==2.2.2
huggingface-hub==0.20.2
faiss-cpu==1.13.2
numpy==1.26.4
pandas==1.5.3
streamlit==1.28.1
python-dotenv==1.0.0
pydantic==2.5.0
pydantic-settings==2.1.0

# PDF processing
PyPDF2==3.0.1
